Fix duplicates in node index
The schema migration process did not update the search index in any way.
The only update occured just before the async schema migration was triggered.
I added the needed calls in the migration handler and added a test.
There is no need to handle micronodes since micronode data is not yet searchable.